fluss-e2e/fluss-e2e-cli/pyproject.toml
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fluss-e2e-cli"
version = "0.1.0"
description = "Local end-to-end CLI for Apache Fluss."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.0",
  "docker>=7.0",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "requests>=2.28",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
fluss-e2e = "fluss_e2e.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fluss_e2e*"]
